refactor:cli/move-related-into-omnifile
