+++ cmd/kjspkg/c_list.go
@@ -46,3 +46,26 @@ func (c *CList) Run(ctx *Context) error {
 	}
 	return nil
 }
+
+type CListall struct {
+	Search string `help:"The search query"`
+}
+
+func (c *CListall) Run(ctx *Context) error {
+	cmd := CList{
+		All:    true,
+		Search: c.Search,
+	}
+	return cmd.Run(ctx)
+}
+
+type CSearch struct {
+	Query string `arg:"" help:"The search query"`
+}
+
+func (c *CSearch) Run(ctx *Context) error {
+	cmd := CListall{
+		Search: c.Query,
+	}
+	return cmd.Run(ctx)
+}
